Add authHierarchyExistence util
- instead of a large code block inside of run function
- remove runningType immediateRun code from run function
