+++ src/schedule.js
@@ -1,6 +1,5 @@
 const dataSync = require('./util/data_synchronizeData');
 const fn = require('./config/fieldNames');
-const p = require('./config/paths');
 const diffsHandler = require('./util/diffsHandler');
 const { sendLog, logLevel } = require('./util/logger');
 const Auth = require('./auth/auth');
@@ -8,6 +7,7 @@ const schedule = require('node-schedule');
 const PromiseAllWithFails = require('./util/generalUtils/promiseAllWithFails');
 const logDetails = require('./util/logDetails');
 const connectToRedis = require('./util/generalUtils/connectToRedis');
+const authHierarchyExistence = require('./util/generalUtils/authHierarchyExistence');
 const express = require("express")
 const moment = require('moment');
 const getRawData = require('./util/getRawData');
@@ -37,46 +37,28 @@ app.post("/immediateRun", async (req, res) => {
 
 app.listen(port, () => console.log("immediateRun server run on port:" + port))
 
-const run = async (runnigType, dataSource, identifiersArray) => {
+const run = async runnigType => {
     try {
         const redis = await connectToRedis();
 
         // check if the root hierarchy exist and adding it if not
-        await Auth.axiosKartoffel.get(p(encodeURIComponent(fn.rootHierarchy.ourCompany)).KARTOFFEL_HIERARCHY_EXISTENCE_CHECKING_BY_DISPLAYNAME_API)
-            .then((result) => {
-                sendLog(logLevel.info, logDetails.info.INF_ROOT_EXSIST, result.data.name);
-            })
-            .catch(async () => {
-                await Auth.axiosKartoffel.post(p().KARTOFFEL_ADDGROUP_API, { name: fn.rootHierarchy.ourCompany })
-                    .then((result) => {
-                        sendLog(logLevel.info, logDetails.info.INF_ADD_ROOT, result.data.name);
-                    })
-                    .catch((err) => {
-                        let errorMessage = (err.response) ? err.response.data.message : err.message;
-                        sendLog(logLevel.error, logDetails.error.ERR_ADD_ROOT, errorMessage);
-                    })
-            });
-
+        authHierarchyExistence();
         // get the new json from aka & save him on the server
         let aka_data = await dataSync(fn.dataSources.aka, runnigType);
 
-        if(runnigType == fn.runnigTypes.ImmediateRun) {
+        await PromiseAllWithFails([
+            GetDataAndProcess(fn.dataSources.es, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.ads, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.adNN, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.lmn, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.mdn, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.mm, aka_data, runnigType, dataSync),
+            GetDataAndProcess(fn.dataSources.city, aka_data, runnigType, dataSync),
+        ]);
+        
+        // Due performence reasons aka flow is run by itself, after the other flows
+        await GetDataAndProcess(fn.dataSources.aka, aka_data);
 
-            await PromiseAllWithFails([
-                GetDataAndProcess(dataSource, aka_data, runnigType, dataSync),
-            ]);
-        } else {
-            await PromiseAllWithFails([
-                GetDataAndProcess(fn.dataSources.aka, aka_data),
-                GetDataAndProcess(fn.dataSources.es, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.ads, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.adNN, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.lmn, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.mdn, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.mm, aka_data, runnigType, dataSync),
-                GetDataAndProcess(fn.dataSources.city, aka_data, runnigType, dataSync),
-            ]);
-        }
         if (redis && redis.status === 'ready') redis.quit();
     } catch (err) {
         sendLog(logLevel.error, logDetails.error.ERR_UN_HANDLED_ERROR, runnigType, JSON.stringify(err));
@@ -100,30 +82,17 @@ const runImmediate = async (dataSource, identifiersArray) => {
         const redis = await connectToRedis();
 
         // check if the root hierarchy exist and adding it if not
-        await Auth.axiosKartoffel.get(p(encodeURIComponent(fn.rootHierarchy.ourCompany)).KARTOFFEL_HIERARCHY_EXISTENCE_CHECKING_BY_DISPLAYNAME_API)
-            .then((result) => {
-                sendLog(logLevel.info, logDetails.info.INF_ROOT_EXSIST, result.data.name);
-            })
-            .catch(async () => {
-                await Auth.axiosKartoffel.post(p().KARTOFFEL_ADDGROUP_API, { name: fn.rootHierarchy.ourCompany })
-                    .then((result) => {
-                        sendLog(logLevel.info, logDetails.info.INF_ADD_ROOT, result.data.name);
-                    })
-                    .catch((err) => {
-                        let errorMessage = (err.response) ? err.response.data.message : err.message;
-                        sendLog(logLevel.error, logDetails.error.ERR_ADD_ROOT, errorMessage);
-                    })
-            });
+        authHierarchyExistence();
 
-            const dateAndTime = moment(new Date()).format("DD.MM.YYYY__HH.mm");
-            let aka_data = await getRawData(fn.dataSources.aka, fn.runnigTypes.ImmediateRun, dateAndTime);
-            let data = {
-                updated: []
-            };
-            data.added = await getRawData(dataSource, fn.runnigTypes.ImmediateRun, dateAndTime, identifiersArray);
-            await diffsHandler(data, dataSource, aka_data);
+        const dateAndTime = moment(new Date()).format("DD.MM.YYYY__HH.mm");
+        let aka_data = await getRawData(fn.dataSources.aka, fn.runnigTypes.ImmediateRun, dateAndTime);
+        let data = {
+            updated: []
+        };
+        data.added = await getRawData(dataSource, fn.runnigTypes.ImmediateRun, dateAndTime, identifiersArray);
+        await diffsHandler(data, dataSource, aka_data);
 
-
+        if (redis && redis.status === 'ready') redis.quit();
 
     } catch (err) {
         sendLog(logLevel.error, logDetails.error.ERR_UN_HANDLED_ERROR, fn.runnigTypes.ImmediateRun, JSON.stringify(err));
